agginta relazioni prima parte

+++ app/Http/Controllers/Admin/PostController.php
@@ -6,6 +6,9 @@ use App\Http\Controllers\Controller;
 use App\Post;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Routing\Route;
+
 
 class PostController extends Controller
 {
@@ -16,14 +19,21 @@ class PostController extends Controller
 
     private function getValidators($model) {
         return [
+            // 'user_id'   => 'required|exists:App\User,id',
             'title'     => 'required|max:100',
-            'slug' => [
+            'slug'      => [
                 'required',
                 Rule::unique('posts')->ignore($model),
                 'max:100'
             ],
             'content'   => 'required'
         ];
+    }
+
+    public function myindex() {
+        $posts = Post::where('user_id', Auth::user()->id)->paginate(50);
+
+        return view('admin.posts.index', compact('posts'));
     }
 
     /**
@@ -58,7 +68,10 @@ class PostController extends Controller
     {
         $request->validate($this->getValidators(null));
 
-        $post = Post::create($request->all());
+        $formData = $request->all() + [
+            'user_id' => Auth::user()->id
+        ];
+        $post = Post::create($formData);
 
         return redirect()->route('admin.posts.show', $post->slug);
     }
@@ -82,6 +95,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if (Auth::user()->id !== $post->user_id) abort(403);
         return view('admin.posts.edit', compact('post'));
     }
 
@@ -94,6 +108,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if (Auth::user()->id !== $post->user_id) abort(403);
+
         $request->validate($this->getValidators($post));
 
         $post->update($request->all());
@@ -109,8 +125,13 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Auth::user()->id !== $post->user_id) abort(403);
+
         $post->delete();
 
-        return redirect()->route('admin.posts.index');
+        if (url()->previous() === route('admin.posts.edit', $post->slug)) {
+            return redirect()->route('admin.home')->with('status', "Post $post->title deleted");;
+        }
+        return redirect(url()->previous())->with('status', "Post $post->title deleted");;
     }
 }
